Fix another error

+++ db/migrate/20230111211952_fix_tags_table.rb
@@ -1,8 +1,13 @@
 class FixTagsTable < ActiveRecord::Migration[7.0]
     def change
-      add_column :tags, :name, :string
-      Tag.update_all(name: '-')
-      change_column_null :tags, :name, false
-      remove_column :tags, :question_id, :integer
+      if column_exists?(:tags,: :name)
+        Tag.update_all(name: '-')
+        change_column_null :tags, :name, false
+      end
+
+      if column_exists?(:tags,: :question_id)
+        remove_column :tags, :question_id, :integer
+      end
+    
     end  
 end
